Publish showWritePermissionSettings as part of public API
showWritePermissionSettings, a method that starts the intent to achieve
MANAGE_WRITE_SETTINGS, was present in the plugin already from the beginning
in WifiApManager class but was not part of the API and therefore not usable
directly.

This MR publishes it so it can be effectively used by code that requires
such permissions (ie. configuration of AP on devices) without the need
to have double code.