Gracefully handle permission requests (#169)
Existing bug:
When a permission needs to be requested (e.g. fine location to scan the wifi networks), we do not wait until the user has finished handling the request before making the API call, resulting in the call failing since the permission hasn't been granted yet.
Once the permission had been granted, this issue would disappear.

Fix:
The result of the permission request is given in the [onRequestPermissionsResult] callback. Only call the API in this callback once the request has been answered by the user. The method success callback and required arguments are saved as field. Only one request can be handled at a time.
When the permission is already granted, the previous call path is kept.
The core functionality of the function with permission requests has been extracted into separate functions.

Co-authored-by: Pierre Hallot <pierre.hallot@flarm.com>